Clean up unused service, add interfaces for others
