--- backend/src/main/java/com/github/dreamteam/services/ScoreSearchService.java
@@ -1,8 +1,0 @@
-package com.github.dreamteam.services;
-
-import java.util.Collection;
-import org.bson.Document;
-
-public interface ScoreSearchService {
-    Collection<Document> searchScores(String projectId);
-}
